Add github action workflows for building production docker images
I've copied the Dockerfile into Dockerfile.prod so that the development
environment doesn't have to be the same as the production environment
(allowing for automated tests etc. in development).

On all pushes/merges into the main branch, a docker image will be pushed
to mlibrary/patron-account-unstable.

On all version releases, a docker image will be pushed to
mlibrary/patron-account.